Step 11: Separate Actors from Actions
Now we have Actors and Actions as distinct collections. They have the
exact same behavior, but we want to give them clear names. Hence the
little bit of inheritance.

Incidents had to change to allow it to accept actors and actions,
instead of one already-built incident hash. And the build_incidents
method has gotten a little nastier as a result.